# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for xaos.mobi bot",
        "title": "Go xaos.mobi bot",
        "contact": {},
        "version": "0.0.0"
    },
    "host": "localhost:5504",
    "basePath": "/",
    "paths": {
        "/account/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get all game accounts",
                "operationId": "get-all-accounts",
                "responses": {}
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Add new game account",
                "operationId": "add-new-game-account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of account`s owner",
                        "name": "owner",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/account/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Get game account by id",
                "operationId": "get-account-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/refresh": {
            "get": {
                "description": "restart task manager in order to add new accounts",
                "tags": [
                    "General"
                ],
                "summary": "Restart task manager",
                "operationId": "restart-task-manager",
                "responses": {}
            }
        },
        "/task/": {
            "get": {
                "description": "get all statuses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get all statuses",
                "operationId": "get-all-status",
                "responses": {}
            }
        },
        "/task/{id}": {
            "get": {
                "description": "get status by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get status by id",
                "operationId": "get-status-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "get status by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update status by id",
                "operationId": "get-status-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "new config",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.Status"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "task.Status": {
            "description": "Model with information about tasks and their parameters",
            "type": "object",
            "properties": {
                "arena_farming": {
                    "type": "boolean"
                },
                "arena_use_energy_cans": {
                    "type": "boolean"
                }
            }
        }
    }
}